Fix scroll locking for fast scrolling
The previous scroll locking only worked when scrolling single lines
(i.e. with the arrow keys or `j` or `k`). This commit fixes the scroll
locking to also be respected when using page- or half-page-scrolling.
It does so by checking whether an update would jump across the center of
the viewport and (if it does) corrects the top_line to be at the correct
position relative to the next current line position.