Initial configuration validation layout
The `Config.validate` command shall be used to validate the
configuration at runtime. This function is NOT called in `set_config`
directly to allow simpler testing. Instead, the `__main__` method should
call this function and exit execution if an error occurs.
Additionally, unittests shall be used to test the detection of possible
misconfiguration.