Implement basic `select` functionality
Upon triggering the select command the current entry under the cursor
will be highlighted with the respective color pair (configurable through
its new option). The highlighting is achieved through substitution of
the viewport's buffer contents replacing the label (plus two spaces to
ensure it's in the first column) with the label (plus ANSI color codes).
Since `select` is a toggle command, re-selecting an entry reverts the
operation.