Disable Select command in search view
Due to the already existing highlighting in the search view, the select
highlighting is not as straight forward to implement in this mode. This
will hopefully simplify in the future once we have a full command
history and can simply revert to the previous viewport state.

However, for now this works fine with this limitation.