WIP: Initial specification of an InputBuffer
The idea here is to wrap `sys.stdin` with an `InputBuffer` which allows
controllable access to the user input. Because `sys.stdin` will usually
be accompanied with some prompt text on `stdout`, the `InputBuffer`
object gets linked to a `TextBuffer` in order to be able to display the
prompt in a popup in the TUI.

The actual logic of handling the input is still missing. But the
previous commit prepared the use of the `prompt_handler` function for
exactly this purpose.