Integrate git into InitCommand
The `init` command now takes the optional `--git` argument which
initializes a git repository in which the database file will be tracked.
Edge cases of existing or missing database files as well as git
repositories are caught and the user is informed, respectively.

This commit also begins removing the functionality of the (deprecated)
--force argument.