Explicitly track Frame mode in State
Rather than relying on a basic distinction of the list and other modes,
we now explicitly track the Frame contents for each possible output
style. The Mode is a simple Enum to allow easy addition later on.
However, the mode values are set as strings during the commands to keep
them as separated as possible from the TUI implementation.