WIP: Configuration rehaul
This is the first step to changing CoBib's configuration to Python.
Based on [this Stackoverlow answer][1], the new `Config` is a subclass
of Python's `dict`. Besides the usual field access, the elements can
also be set and accessed via attributes.

Note: as this is a WIP and standlone commit, no unittest is likely to
pass for this specific commit.

[1]: https://stackoverflow.com/a/3031270