Add support for legacy INI-style configuration
To ease the transition for users from the INI-style to Python-style
configuration, the legacy kind will be respected until 1.1.2022.
However, to provide the users more incentive to switch, new options will
not be integrated into the legacy configuration (in fact, some of the
things that are planned will not be possible in the old configuration
style which is the main reason for the re-design).