Make config.__setattr__ smarter
With the previous implementation where `__setattr__` was set to
`__setitem__`, property setters where essentially rendered useless,
because `__setattr__` always took priority.
Trying to avoid implementing a hard-coded list of exceptions for the
keys in the `__setitem__` method, I came across [this solution][1] on
Stackoverflow which checks whether the key is a `property` object. If
so, its setter method will be used. Otherwise, the `__setitem__` method
is called.

This permits the definition of the internal `_bibliography` attribute
which can be exposed as a `property` which also supports safe deletion
in the form of clearing the `OrderedDict`.