Fully exclude internal attributes from automagic Config creation
Thinking a little bit more about the issue with the `__wrapped__`
attribute, I realized that the problem had actually to do with the
automagic population of non-existing attributes with an empty Config
object.
Similarly to the specific case caught during `__setattr__`, this commit
introduces an exception in the `__getattr__` case to exclude internal
attributes (those starting with `__`) from this behavior.
Thus, this fixes the `__wrapped__` attribute issue in a more general
fashion and (hopefully) obviates further issues that could be caused by
such problems somewhere in Python's depths...