Fix forked process weirdness
The unittests have always been suffering from some weirdness with
respect to the forking nature of the `pyte`-based tests. With the
addition of the `pytest-cov` plugin this causes frequent CI failures.
Sending a `SIGTERM` signal to the forked child process appears to fixes
the weirdness locally, let's hope it does the same on the CI...